+++ src/pages/Assets.vue
@@ -52,16 +52,26 @@
         </router-link>
       </q-td>
 
+      <q-td slot="body-cell-description" slot-scope="props" :props="props"
+      >
+        <pxs-markdown simple hide-empty :text="props.value"/>
+      </q-td>
+
     </q-table>
   </q-page>
 </template>
 
 <script>
   import allAssetsList from '../graphql/assets.gql'
+  import PxsMarkdown from 'components/pxs-markdown'
 
   const debug = false
 
   export default {
+    components: {
+      PxsMarkdown,
+    },
+
     data() {
       return {
         allAssetsList: [],