+++ src/pages/Missions.vue
@@ -67,6 +67,11 @@
         </router-link>
       </q-td>
 
+      <q-td slot="body-cell-description" slot-scope="props" :props="props"
+      >
+        <pxs-markdown simple hide-empty :text="props.value"/>
+      </q-td>
+
       <q-td slot="body-cell-missionStatus" slot-scope="props" :props="props"
             style="width:5px"
       >
@@ -78,14 +83,16 @@
 </template>
 
 <script>
-  import MissionNewButton from 'components/mission-new-button'
   import allMissionsList from '../graphql/missions.gql'
+  import MissionNewButton from 'components/mission-new-button'
+  import PxsMarkdown from 'components/pxs-markdown'
 
   const debug = false
 
   export default {
     components: {
-      MissionNewButton
+      MissionNewButton,
+      PxsMarkdown,
     },
 
     data() {